controller: Added controller for user route
Separating user route into its own controller will help modularity and easy
testing of business logic.
